feat: Add functionality to calculate the sum of a range of numbers and check if a number is prime
This commit adds two new functions to the code. The first function, `calculate_sum`, takes two numbers as input and calculates the sum of the numbers in the range between them. The second function, `is_prime`, checks if a given number is prime. The code now allows the user to input a range of numbers and calculates the sum of the range. It also checks if a number is prime and provides the corresponding output.

Note: The commit message has been generated based on the provided code changes and does not include any meta information.